requirements.txt
streamlit
graphviz
google-generativeai
python-dotenv
pathlib
ipykernel
seaborn